Move new code out of legacy. De-engineer PathResourcePack.
